orchestrator/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.83.0
anyio==4.12.1
asgiref==3.11.1
astroid==4.0.4
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
dill==0.4.1
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.129.1
filetype==1.2.0
frozenlist==1.8.0
google-api-core==2.30.0
google-auth==2.48.0
google-genai==1.64.0
googleapis-common-protos==1.72.0
grpcio==1.78.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
isort==8.0.0
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-anthropic==1.3.3
langchain-core==1.2.14
langchain-google-genai==4.2.1
langgraph==1.0.9
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.8
langsmith==0.7.6
mccabe==0.7.0
multidict==6.7.1
neo4j==6.1.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-prometheus==0.60b1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-asgi==0.60b1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
platformdirs==4.9.2
pluggy==1.6.0
prometheus_client==0.24.1
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
PyJWT==2.11.0
pydantic_core==2.41.5
Pygments==2.19.2
pylint==4.0.5
pytest==9.0.2
pytest-asyncio==1.3.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
starlette==0.52.1
tenacity==9.1.4
tree-sitter==0.25.2
tree-sitter-go==0.25.0
tomlkit==0.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
websockets==15.0.1
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
